// src/features/payments/enums/paymentsRejectionReason.enum.ts
export enum PaymentRejectionReasonType {
    SALDO_INSUFICIENTE = 'SALDO_INSUFICIENTE',
    VALOR_ACIMA_LIMITE = 'VALOR_ACIMA_LIMITE',
    VALOR_INVALIDO = 'VALOR_INVALIDO',
    COBRANCA_INVALIDA = 'COBRANCA_INVALIDA',
    NAO_INFORMADO = 'NAO_INFORMADO',
    PAGAMENTO_DIVERGENTE_CONSENTIMENTO = 'PAGAMENTO_DIVERGENTE_CONSENTIMENTO',
    DETALHE_PAGAMENTO_INVALIDO = 'DETALHE_PAGAMENTO_INVALIDO',
    PAGAMENTO_RECUSADO_DETENTORA = 'PAGAMENTO_RECUSADO_DETENTORA',
    PAGAMENTO_RECUSADO_SPI = 'PAGAMENTO_RECUSADO_SPI',
    FALHA_INFRAESTRUTURA = 'FALHA_INFRAESTRUTURA',
    FALHA_INFRAESTRUTURA_SPI = 'FALHA_INFRAESTRUTURA_SPI',
    FALHA_INFRAESTRUTURA_DICT = 'FALHA_INFRAESTRUTURA_DICT',
    FALHA_INFRAESTRUTURA_ICP = 'FALHA_INFRAESTRUTURA_ICP',
    FALHA_INFRAESTRUTURA_PSP_RECEBEDOR = 'FALHA_INFRAESTRUTURA_PSP_RECEBEDOR',
    FALHA_INFRAESTRUTURA_DETENTORA = 'FALHA_INFRAESTRUTURA_DETENTORA',
    CONTAS_ORIGEM_DESTINO_IGUAIS = 'CONTAS_ORIGEM_DESTINO_IGUAIS',
    PAGAMENTO_NAO_PERMITE_CANCELAMENTO = 'PAGAMENTO_NAO_PERMITE_CANCELAMENTO',
}